Fixed classes Item and Character on record. Moved the files to a separate folder
