--- src/main/java/faang/school/godbless/stream_split/Item.java
@@ -1,15 +1,0 @@
-package faang.school.godbless.stream_split;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@AllArgsConstructor
-@ToString
-public class Item {
-
-    private String name;
-
-    private int value;
-}
