--- src/main/java/faang/school/godbless/stream_split/Main.java
@@ -1,21 +1,0 @@
-package faang.school.godbless.stream_split;
-
-public class Main {
-
-    public static void main(String[] args) {
-        Character frodo = new Character("Frodo");
-        Item ring = new Item("The One Ring", 1000);
-
-        InventoryManager manager = new InventoryManager();
-
-// Добавляем предмет в инвентарь
-        manager.addItem(frodo, ring, (item) -> System.out.println(item + " was added to the inventory."));
-
-// Удаляем предмет из инвентаря
-        manager.removeItem(frodo, (item) -> item.getName().equals("The One Ring"));
-
-// Обновляем предмет в инвентаре
-        manager.updateItem(frodo, (item) -> item.getName().equals("The One Ring"), (item) -> new Item(item.getName(), item.getValue() * 2));
-        System.out.println(frodo);
-    }
-}
