+++ src/main/java/faang/school/godbless/stream_split/rpg_game/Character.java
@@ -1,0 +1,6 @@
+package faang.school.godbless.stream_split.rpg_game;
+
+import java.util.List;
+
+public record Character(String name, List<Item> inventory) {
+}
