+++ src/main/java/faang/school/godbless/stream_split/rpg_game/Item.java
@@ -1,0 +1,3 @@
+package faang.school.godbless.stream_split.rpg_game;
+
+public record Item(String name, int value) {}
